fix(types): #919——Bar.volume i64→f64（小数成交量入库不截断），全消费面跟随
- Bar.volume: f64 + derive 去 Eq（f64 无 Eq；Bar 无 HashMap 键用途，PartialEq 足够——对拍场景全在测试）
- 截断点全修：bar_adapter（nb.volume as f64）、data.rs/tick.rs loader、inclusion 累加（+ 替 saturating_add）、backtest_engine Quantity::new、40+ 探针 bin/测试字面量跟随
- 修复见证断言翻正：qty=0.00003 / 0.00279 直存非零（旧 i64 口径截 0）
- 全绿 2825/0 + FMT_OK；e2e BTC 两月窗订单 2754/持仓 177 与改前逐位一致（整数成交量数据无损）
- 白名单七夹具点随 fmt 重排再重锚（+2 漂移）

+++ rust/src/theta_v0/backtest/data.rs
@@ -319,7 +319,7 @@ pub fn load_symbol(
             high: hq,
             low: lq,
             close: cq,
-            volume: v as i64,
+            volume: v as f64, // #919：f64 直存
             untradable,
         });
     }
@@ -463,7 +463,7 @@ mod tests {
                     high: 1,
                     low: 1,
                     close: 1,
-                    volume: 1,
+                    volume: 1.0,
                     untradable: false,
                 })
                 .collect(),
@@ -544,7 +544,7 @@ mod tests {
             high: 1,
             low: 1,
             close: 1,
-            volume: 1,
+            volume: 1.0,
             untradable,
         };
         let ds = Dataset {
